Add files via upload
Adding these two files which are modifications of testIndex.html and holdingsTable.js to add 'Get Token Balance' button, which onclick will read token balance from metamask and display token balance in a table in the webpage.  In order for code to work, metamask must be logged into and connected to the website running the code.